api/requirements.txt
supabase==2.0.0
python-dotenv==1.0.0
pydantic==2.5.3